Pulled the SSL config out into a shared class.
